Rename package to deptemplateapi
