+++ pkg/api/deploymentapi/deptemplateapi/delete.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package deploymenttemplateapi
+package deptemplateapi
 
 import (
 	"errors"
@@ -23,34 +23,28 @@ import (
 	"github.com/elastic/cloud-sdk-go/pkg/api"
 	"github.com/elastic/cloud-sdk-go/pkg/api/apierror"
 	"github.com/elastic/cloud-sdk-go/pkg/client/deployment_templates"
-	"github.com/elastic/cloud-sdk-go/pkg/models"
 	"github.com/elastic/cloud-sdk-go/pkg/multierror"
 	"github.com/elastic/cloud-sdk-go/pkg/util/ec"
 )
 
-// UpdateParams is consumed by the Update function.
-type UpdateParams struct {
+// DeleteParams is consumed by the Delete function.
+type DeleteParams struct {
 	*api.API
 
-	Region     string
 	TemplateID string
-	Request    *models.DeploymentTemplateRequestBody
+	Region     string
 }
 
-// Validate ensures the parameters are usable by Update.
-func (params UpdateParams) Validate() error {
-	var merr = multierror.NewPrefixed("invalid deployment template update params")
+// Validate ensures the parameters are usable by Delete.
+func (params DeleteParams) Validate() error {
+	var merr = multierror.NewPrefixed("invalid deployment template delete params")
 
 	if params.API == nil {
 		merr = merr.Append(apierror.ErrMissingAPI)
 	}
 
-	if params.Request == nil {
-		merr = merr.Append(errors.New("required template request definition not provided"))
-	}
-
 	if params.TemplateID == "" {
-		merr = merr.Append(errors.New("required template id not provided"))
+		merr = merr.Append(errors.New("required template ID not provided"))
 	}
 
 	if err := ec.RequireRegionSet(params.Region); err != nil {
@@ -60,20 +54,18 @@ func (params UpdateParams) Validate() error {
 	return merr.ErrorOrNil()
 }
 
-// Update updates an existing deployment template from a definition.
-func Update(params UpdateParams) error {
+// Delete removes an existing deployment temeplate.
+func Delete(params DeleteParams) error {
 	if err := params.Validate(); err != nil {
 		return err
 	}
 
-	_, _, err := params.V1API.DeploymentTemplates.SetDeploymentTemplateV2(
-		deployment_templates.NewSetDeploymentTemplateV2Params().
-			WithTemplateID(params.TemplateID).
-			WithCreateOnly(ec.Bool(false)).
-			WithRegion(params.Region).
-			WithBody(params.Request),
-		params.AuthWriter,
+	return api.ReturnErrOnly(
+		params.V1API.DeploymentTemplates.DeleteDeploymentTemplateV2(
+			deployment_templates.NewDeleteDeploymentTemplateV2Params().
+				WithRegion(params.Region).
+				WithTemplateID(params.TemplateID),
+			params.AuthWriter,
+		),
 	)
-
-	return apierror.Unwrap(err)
 }
