+++ pkg/api/deploymentapi/deptemplateapi/delete_test.go
@@ -1,0 +1,96 @@
+// Licensed to Elasticsearch B.V. under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. Elasticsearch B.V. licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package deptemplateapi
+
+import (
+	"errors"
+	"net/url"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+
+	"github.com/elastic/cloud-sdk-go/pkg/api"
+	"github.com/elastic/cloud-sdk-go/pkg/api/mock"
+	"github.com/elastic/cloud-sdk-go/pkg/multierror"
+)
+
+func TestDelete(t *testing.T) {
+	type args struct {
+		params DeleteParams
+	}
+	tests := []struct {
+		name string
+		args args
+		err  error
+	}{
+		{
+			name: "fails due to parameter validation",
+			err: multierror.NewPrefixed("invalid deployment template delete params",
+				errors.New("api reference is required for the operation"),
+				errors.New("required template ID not provided"),
+				errors.New("region not specified and is required for this operation"),
+			),
+		},
+		{
+			name: "succeeds",
+			args: args{params: DeleteParams{
+				Region:     "us-east-1",
+				TemplateID: "some-id",
+				API: api.NewMock(mock.New200ResponseAssertion(
+					&mock.RequestAssertion{
+						Header: api.DefaultWriteMockHeaders,
+						Method: "DELETE",
+						Host:   api.DefaultMockHost,
+						Path:   "/api/v1/deployments/templates/some-id",
+						Query: url.Values{
+							"region": []string{"us-east-1"},
+						},
+					},
+					mock.NewStringBody(`{}`),
+				)),
+			}},
+		},
+		{
+			name: "fails on API error",
+			args: args{params: DeleteParams{
+				Region:     "us-east-1",
+				TemplateID: "some-id",
+				API: api.NewMock(mock.New500ResponseAssertion(
+					&mock.RequestAssertion{
+						Header: api.DefaultWriteMockHeaders,
+						Method: "DELETE",
+						Host:   api.DefaultMockHost,
+						Path:   "/api/v1/deployments/templates/some-id",
+						Query: url.Values{
+							"region": []string{"us-east-1"},
+						},
+					},
+					mock.SampleInternalError().Response.Body,
+				)),
+			}},
+			err: mock.MultierrorInternalError,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if err := Delete(tt.args.params); !assert.Equal(t, tt.err, err) {
+				t.Error(err)
+			}
+		})
+	}
+}
