+++ pkg/api/deploymentapi/deptemplateapi/get.go
@@ -15,11 +15,10 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package deploymenttemplateapi
+package deptemplateapi
 
 import (
-	"fmt"
-	"strings"
+	"errors"
 
 	"github.com/elastic/cloud-sdk-go/pkg/api"
 	"github.com/elastic/cloud-sdk-go/pkg/api/apierror"
@@ -29,32 +28,27 @@ import (
 	"github.com/elastic/cloud-sdk-go/pkg/util/ec"
 )
 
-// ListParams is consumed by the List function.
-type ListParams struct {
+// GetParams is consumed by the Get function.
+type GetParams struct {
 	*api.API
 
-	MetadataFilter string
-	Region         string
-	StackVersion   string
-
-	ShowHidden bool
+	TemplateID   string
+	Region       string
+	StackVersion string
 
 	HideInstanceConfigurations bool
 }
 
-// Validate ensures the parameters are usable by List.
-func (params ListParams) Validate() error {
-	var merr = multierror.NewPrefixed("invalid deployment template list params")
+// Validate ensures the parameters are usable by Get.
+func (params GetParams) Validate() error {
+	var merr = multierror.NewPrefixed("invalid deployment template get params")
 
 	if params.API == nil {
 		merr = merr.Append(apierror.ErrMissingAPI)
 	}
 
-	if params.MetadataFilter != "" && len(strings.Split(params.MetadataFilter, ":")) != 2 {
-		merr = merr.Append(fmt.Errorf(
-			`invalid metadata filter "%s", must be formatted in the form of (key:value)`,
-			params.MetadataFilter,
-		))
+	if params.TemplateID == "" {
+		merr = merr.Append(errors.New("required template ID not provided"))
 	}
 
 	if err := ec.RequireRegionSet(params.Region); err != nil {
@@ -64,14 +58,14 @@ func (params ListParams) Validate() error {
 	return merr.ErrorOrNil()
 }
 
-// List returns a list of the available deployment templates.
-func List(params ListParams) ([]*models.DeploymentTemplateInfoV2, error) {
+// Get returns the specified deployment template.
+func Get(params GetParams) (*models.DeploymentTemplateInfoV2, error) {
 	if err := params.Validate(); err != nil {
 		return nil, err
 	}
 
-	res, err := params.V1API.DeploymentTemplates.GetDeploymentTemplatesV2(
-		listParams(params), params.AuthWriter,
+	res, err := params.V1API.DeploymentTemplates.GetDeploymentTemplateV2(
+		getParams(params), params.AuthWriter,
 	)
 
 	if err != nil {
@@ -81,21 +75,17 @@ func List(params ListParams) ([]*models.DeploymentTemplateInfoV2, error) {
 	return res.Payload, nil
 }
 
-// listParams creates the API params from the intermediary params struct.
+// getParams creates the API params from the intermediary params struct.
 // This is required for fields which are *<type> instead of directly using
 // With<Param>() builders which don't work well for pointer of types on default
 // values (i.e. "" => <string>, 0 => <int>, false => <bool>).
 // and the function makes it easier to use since all the logic to create and
 // set params is contained here.
-func listParams(params ListParams) *deployment_templates.GetDeploymentTemplatesV2Params {
-	var apiParams = deployment_templates.NewGetDeploymentTemplatesV2Params().
+func getParams(params GetParams) *deployment_templates.GetDeploymentTemplateV2Params {
+	var apiParams = deployment_templates.NewGetDeploymentTemplateV2Params().
 		WithShowInstanceConfigurations(ec.Bool(!params.HideInstanceConfigurations)).
-		WithShowHidden(&params.ShowHidden).
-		WithRegion(params.Region)
-
-	if params.MetadataFilter != "" {
-		apiParams.SetMetadata(&params.MetadataFilter)
-	}
+		WithRegion(params.Region).
+		WithTemplateID(params.TemplateID)
 
 	if params.StackVersion != "" {
 		apiParams.SetStackVersion(&params.StackVersion)
