+++ pkg/api/deploymentapi/deptemplateapi/get_test.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package deploymenttemplateapi
+package deptemplateapi
 
 import (
 	"encoding/json"
@@ -32,126 +32,91 @@ import (
 	"github.com/elastic/cloud-sdk-go/pkg/multierror"
 )
 
-func TestList(t *testing.T) {
-	listRawResp, err := ioutil.ReadFile("./testdata/list.json")
+func TestGet(t *testing.T) {
+	getRawResp, err := ioutil.ReadFile("./testdata/get.json")
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	var succeedResp []*models.DeploymentTemplateInfoV2
-	if err := json.Unmarshal(listRawResp, &succeedResp); err != nil {
+	var succeedResp *models.DeploymentTemplateInfoV2
+	if err := json.Unmarshal(getRawResp, &succeedResp); err != nil {
 		t.Fatal(err)
 	}
 	type args struct {
-		params ListParams
+		params GetParams
 	}
 	tests := []struct {
 		name string
 		args args
-		want []*models.DeploymentTemplateInfoV2
+		want *models.DeploymentTemplateInfoV2
 		err  error
 	}{
 		{
 			name: "fails due to parameter validation",
-			err: multierror.NewPrefixed("invalid deployment template list params",
+			err: multierror.NewPrefixed("invalid deployment template get params",
 				errors.New("api reference is required for the operation"),
-				errors.New("region not specified and is required for this operation"),
-			),
-		},
-		{
-			name: "fails due to parameter invalid metadata filter",
-			args: args{params: ListParams{
-				MetadataFilter: "somewrongful value",
-			}},
-			err: multierror.NewPrefixed("invalid deployment template list params",
-				errors.New("api reference is required for the operation"),
-				errors.New(`invalid metadata filter "somewrongful value", must be formatted in the form of (key:value)`),
+				errors.New("required template ID not provided"),
 				errors.New("region not specified and is required for this operation"),
 			),
 		},
 		{
 			name: "succeeds",
-			args: args{params: ListParams{
-				Region: "us-east-1",
+			args: args{params: GetParams{
+				Region:     "us-east-1",
+				TemplateID: "default",
 				API: api.NewMock(mock.New200ResponseAssertion(
 					&mock.RequestAssertion{
 						Header: api.DefaultReadMockHeaders,
 						Method: "GET",
 						Host:   api.DefaultMockHost,
-						Path:   "/api/v1/deployments/templates",
+						Path:   "/api/v1/deployments/templates/default",
 						Query: url.Values{
 							"region":                       []string{"us-east-1"},
 							"show_instance_configurations": []string{"true"},
-							"show_hidden":                  []string{"false"},
 						},
 					},
-					mock.NewByteBody(listRawResp),
+					mock.NewByteBody(getRawResp),
 				)),
 			}},
 			want: succeedResp,
 		},
 		{
-			name: "succeeds with metadata filter",
-			args: args{params: ListParams{
-				Region:         "us-east-1",
-				MetadataFilter: "parent_solution:stack",
+			name: "succeeds with stack filter",
+			args: args{params: GetParams{
+				Region:       "us-east-1",
+				TemplateID:   "default",
+				StackVersion: "6.8.0",
 				API: api.NewMock(mock.New200ResponseAssertion(
 					&mock.RequestAssertion{
 						Header: api.DefaultReadMockHeaders,
 						Method: "GET",
 						Host:   api.DefaultMockHost,
-						Path:   "/api/v1/deployments/templates",
+						Path:   "/api/v1/deployments/templates/default",
 						Query: url.Values{
-							"metadata":                     []string{"parent_solution:stack"},
 							"region":                       []string{"us-east-1"},
 							"show_instance_configurations": []string{"true"},
-							"show_hidden":                  []string{"false"},
+							"stack_version":                []string{"6.8.0"},
 						},
 					},
-					mock.NewByteBody(listRawResp),
-				)),
-			}},
-			want: succeedResp,
-		},
-		{
-			name: "succeeds with metadata and stack_version filter",
-			args: args{params: ListParams{
-				Region:         "us-east-1",
-				MetadataFilter: "parent_solution:stack",
-				StackVersion:   "6.2.1",
-				API: api.NewMock(mock.New200ResponseAssertion(
-					&mock.RequestAssertion{
-						Header: api.DefaultReadMockHeaders,
-						Method: "GET",
-						Host:   api.DefaultMockHost,
-						Path:   "/api/v1/deployments/templates",
-						Query: url.Values{
-							"metadata":                     []string{"parent_solution:stack"},
-							"region":                       []string{"us-east-1"},
-							"show_instance_configurations": []string{"true"},
-							"show_hidden":                  []string{"false"},
-							"stack_version":                []string{"6.2.1"},
-						},
-					},
-					mock.NewByteBody(listRawResp),
+					mock.NewByteBody(getRawResp),
 				)),
 			}},
 			want: succeedResp,
 		},
 		{
 			name: "fails on API error",
-			args: args{params: ListParams{
-				Region: "us-east-1",
+			args: args{params: GetParams{
+				Region:     "us-east-1",
+				TemplateID: "some-id",
 				API: api.NewMock(mock.New500ResponseAssertion(
 					&mock.RequestAssertion{
 						Header: api.DefaultReadMockHeaders,
 						Method: "GET",
 						Host:   api.DefaultMockHost,
-						Path:   "/api/v1/deployments/templates",
+						Path:   "/api/v1/deployments/templates/some-id",
 						Query: url.Values{
 							"region":                       []string{"us-east-1"},
 							"show_instance_configurations": []string{"true"},
-							"show_hidden":                  []string{"false"},
 						},
 					},
 					mock.SampleInternalError().Response.Body,
@@ -162,7 +127,7 @@ func TestList(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := List(tt.args.params)
+			got, err := Get(tt.args.params)
 			if !assert.Equal(t, tt.err, err) {
 				t.Error(err)
 			}
