+++ pkg/api/deploymentapi/deptemplateapi/testdata/get.json
