+++ pkg/api/deploymentapi/deptemplateapi/testdata/update.json
