+++ pkg/api/deploymentapi/deptemplateapi/update_test.go
@@ -15,10 +15,12 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package deploymenttemplateapi
+package deptemplateapi
 
 import (
+	"encoding/json"
 	"errors"
+	"io/ioutil"
 	"net/url"
 	"testing"
 
@@ -26,12 +28,27 @@ import (
 
 	"github.com/elastic/cloud-sdk-go/pkg/api"
 	"github.com/elastic/cloud-sdk-go/pkg/api/mock"
+	"github.com/elastic/cloud-sdk-go/pkg/models"
 	"github.com/elastic/cloud-sdk-go/pkg/multierror"
 )
 
-func TestDelete(t *testing.T) {
+func TestUpdate(t *testing.T) {
+	updateRawDef, err := ioutil.ReadFile("./testdata/update.json")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	var updateReq *models.DeploymentTemplateRequestBody
+	if err := json.Unmarshal(updateRawDef, &updateReq); err != nil {
+		t.Fatal(err)
+	}
+
+	updateRawBody, err := json.Marshal(updateReq)
+	if err != nil {
+		t.Fatal(err)
+	}
 	type args struct {
-		params DeleteParams
+		params UpdateParams
 	}
 	tests := []struct {
 		name string
@@ -40,45 +57,52 @@ func TestDelete(t *testing.T) {
 	}{
 		{
 			name: "fails due to parameter validation",
-			err: multierror.NewPrefixed("invalid deployment template delete params",
+			err: multierror.NewPrefixed("invalid deployment template update params",
 				errors.New("api reference is required for the operation"),
-				errors.New("required template id not provided"),
+				errors.New("required template request definition not provided"),
+				errors.New("required template ID not provided"),
 				errors.New("region not specified and is required for this operation"),
 			),
 		},
 		{
 			name: "succeeds",
-			args: args{params: DeleteParams{
+			args: args{params: UpdateParams{
+				TemplateID: "my-preset-id",
 				Region:     "us-east-1",
-				TemplateID: "some-id",
-				API: api.NewMock(mock.New200ResponseAssertion(
+				Request:    updateReq,
+				API: api.NewMock(mock.New201ResponseAssertion(
 					&mock.RequestAssertion{
 						Header: api.DefaultWriteMockHeaders,
-						Method: "DELETE",
+						Method: "PUT",
 						Host:   api.DefaultMockHost,
-						Path:   "/api/v1/deployments/templates/some-id",
+						Path:   "/api/v1/deployments/templates/my-preset-id",
 						Query: url.Values{
-							"region": []string{"us-east-1"},
+							"create_only": []string{"false"},
+							"region":      []string{"us-east-1"},
 						},
+						Body: mock.NewStringBody(string(updateRawBody) + "\n"),
 					},
-					mock.NewStringBody(`{}`),
+					mock.NewStringBody(`{"id": "my-preset-id"}`),
 				)),
 			}},
 		},
 		{
 			name: "fails on API error",
-			args: args{params: DeleteParams{
+			args: args{params: UpdateParams{
+				TemplateID: "some",
 				Region:     "us-east-1",
-				TemplateID: "some-id",
+				Request:    &models.DeploymentTemplateRequestBody{},
 				API: api.NewMock(mock.New500ResponseAssertion(
 					&mock.RequestAssertion{
 						Header: api.DefaultWriteMockHeaders,
-						Method: "DELETE",
+						Method: "PUT",
 						Host:   api.DefaultMockHost,
-						Path:   "/api/v1/deployments/templates/some-id",
+						Path:   "/api/v1/deployments/templates/some",
 						Query: url.Values{
-							"region": []string{"us-east-1"},
+							"create_only": []string{"false"},
+							"region":      []string{"us-east-1"},
 						},
+						Body: mock.NewStructBody(&models.DeploymentTemplateRequestBody{}),
 					},
 					mock.SampleInternalError().Response.Body,
 				)),
@@ -88,7 +112,7 @@ func TestDelete(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			if err := Delete(tt.args.params); !assert.Equal(t, tt.err, err) {
+			if err := Update(tt.args.params); !assert.Equal(t, tt.err, err) {
 				t.Error(err)
 			}
 		})
